src/test/java/com/wechat_order/sell/service/impl/SellerServiceImplTest.java
package com.wechat_order.sell.service.impl;

import com.wechat_order.sell.dataobject.SellerInfo;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import static org.junit.Assert.*;

@RunWith(SpringRunner.class)
@SpringBootTest
public class SellerServiceImplTest {

    private static final String openid = "test1";

    @Autowired
    private SellerServiceImpl sellerService;

    @Test
    public void findSellerInfoByOpenId() {
        SellerInfo result = sellerService.findSellerInfoByOpenId(openid);
        Assert.assertEquals(openid,result.getOpenid());
    }
}